Divide the curve into two mods projective_point affine_point
