+++ crates/starknet-types-core/src/curve/mod.rs
@@ -1,0 +1,2 @@
+pub mod affine_point;
+pub mod projective_point;
