+++ crates/starknet-types-core/src/curve/projective_point.rs
@@ -6,11 +6,12 @@ use lambdaworks_math::elliptic_curve::short_weierstrass::point::ShortWeierstrass
 use lambdaworks_math::elliptic_curve::traits::{EllipticCurveError, FromAffine};
 use lambdaworks_math::field::element::FieldElement;
 use lambdaworks_math::field::fields::fft_friendly::stark_252_prime_field::Stark252PrimeField;
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ProjectivePoint(ShortWeierstrassProjectivePoint<StarkCurve>);
+use lambdaworks_math::unsigned_integer::traits::IsUnsignedInteger;
+
+use crate::curve::affine_point::AffinePoint;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AffinePoint(ShortWeierstrassProjectivePoint<StarkCurve>);
+pub struct ProjectivePoint(pub(crate) ShortWeierstrassProjectivePoint<StarkCurve>);
 
 impl ProjectivePoint {
     pub fn new(x: Felt, y: Felt, z: Felt) -> ProjectivePoint {
@@ -60,52 +61,21 @@ impl ops::AddAssign<&ProjectivePoint> for ProjectivePoint {
     }
 }
 
-impl AffinePoint {
-    pub fn new(x: Felt, y: Felt) -> Result<AffinePoint, EllipticCurveError> {
-        Ok(Self(ShortWeierstrassProjectivePoint::from_affine(
-            x.0, y.0,
-        )?))
-    }
+impl ops::Mul<&Felt> for &ProjectivePoint {
+    type Output = ProjectivePoint;
 
-    pub fn identity() -> AffinePoint {
-        Self(ShortWeierstrassProjectivePoint::neutral_element())
-    }
-
-    /// Returns the `x` coordinate of the point.
-    pub fn x(&self) -> &FieldElement<Stark252PrimeField> {
-        self.0.x()
-    }
-
-    /// Returns the `y` coordinate of the point.
-    pub fn y(&self) -> &FieldElement<Stark252PrimeField> {
-        self.0.y()
+    fn mul(self, rhs: &Felt) -> ProjectivePoint {
+        ProjectivePoint(self.0.operate_with_self(rhs.0.representative()))
     }
 }
 
-impl ops::Add<&AffinePoint> for &AffinePoint {
-    type Output = AffinePoint;
+impl<T> ops::Mul<T> for &ProjectivePoint
+where
+    T: IsUnsignedInteger, // Asumiendo que "exponent" es una trait
+{
+    type Output = ProjectivePoint;
 
-    fn add(self, rhs: &AffinePoint) -> AffinePoint {
-        AffinePoint(self.0.operate_with(&rhs.0))
-    }
-}
-
-impl ops::AddAssign<&AffinePoint> for AffinePoint {
-    fn add_assign(&mut self, rhs: &AffinePoint) {
-        self.0 = self.0.operate_with(&rhs.0);
-    }
-}
-
-impl ops::Mul<&Felt> for &AffinePoint {
-    type Output = AffinePoint;
-
-    fn mul(self, rhs: &Felt) -> AffinePoint {
-        AffinePoint(self.0.operate_with_self(rhs.0.representative()))
-    }
-}
-
-impl From<AffinePoint> for ProjectivePoint {
-    fn from(affine_point: AffinePoint) -> ProjectivePoint {
-        ProjectivePoint(affine_point.0)
+    fn mul(self, rhs: T) -> ProjectivePoint {
+        ProjectivePoint(self.0.operate_with_self(rhs))
     }
 }
